Fixed Videos/Text Issue For Sebastian Tab When Screen Shrinks
Fixed the issue where if the windows where to shrink, the videos and text would not align properly and the second video would overlap the footer. Fixed this issue by treating them as columns within a container instead of trying to fix it using css. Also, I removed the extra css file I made since it was unnecessary now.