Fixed PasswordProperty Applying to Email and Username in Create.cshtml
Fixed the issue where when the user enters some value for username and email, the password property (i.e. the words get turned into colored in circles) apply to the username and email. Turns out the issue is that I had used ".PasswordFor" instead of ".EditorFor" which is why it was occurring.